tests/dialog-display.html

<!doctype html>
<meta charset=utf-8>
<title>dialog: display</title>
<script src="./dist/polyfill.js"></script>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
 dialog { position: static }
</style>
<dialog open id=dialog></dialog>
<script>
test(function() {
  assert_equals(getComputedStyle(document.getElementById('dialog')).display, 'block');
});
</script>
